# Conflicts:
#	app/src/main/java/com/jhuoose/foodaholic/api/HerokuAPI.java

+++ app/src/main/java/com/jhuoose/foodaholic/viewmodel/ActivityProfile.java
@@ -5,7 +5,7 @@ public class ActivityProfile {
     private String activityName;
     private String description;
     private int vote;
-    private float money;
+    private double money;
     private String category;
 
     public ActivityProfile() {
@@ -47,11 +47,11 @@ public class ActivityProfile {
         this.vote = vote;
     }
 
-    public float getMoney() {
+    public double getMoney() {
         return money;
     }
 
-    public void setMoney(float money) {
+    public void setMoney(double money) {
         this.money = money;
     }
